+++ composer.json
@@ -7,5 +7,10 @@
       "name": "Matt Miller"
     }
   ],
-  "require": {}
+  "require": {},
+  "autoload": {
+        "psr-4": {
+            "App\\": "app/"
+        }
+    }
 }
