+++ deprecated.php
@@ -1,9 +1,11 @@
 <?php
 
-/*
-Deprecated functions moved to class.
-These functions may be removed in a future version.
-*/
+use App\Helpers;
+
+/**
+ * @deprecated: functions moved to class.
+ * These functions may be removed in a future version.
+ */
 
 // Based on WP core `_deprecated_function()`
 function _ptc_deprecated_warning($function, $version, $replacement = null)
@@ -23,7 +25,7 @@ if (! function_exists('isPasswordProtected')) {
     function isPasswordProtected($post)
     {
         _ptc_deprecated_warning(__FUNCTION__, '1.3.4', 'ProtectTheChildren_Helpers::isPasswordProtected');
-        ProtectTheChildren_Helpers::isPasswordProtected($post);
+        Helpers::isPasswordProtected($post);
     }
 }
 
@@ -31,7 +33,7 @@ if (! function_exists('protectTheChildrenEnabled')) {
     function protectTheChildrenEnabled($post)
     {
         _ptc_deprecated_warning(__FUNCTION__, '1.3.4', 'ProtectTheChildren_Helpers::isEnabled');
-        ProtectTheChildren_Helpers::isEnabled($post);
+        Helpers::isEnabled($post);
     }
 }
 
@@ -39,6 +41,6 @@ if (! function_exists('processPosts')) {
     function processPosts($post)
     {
         _ptc_deprecated_warning(__FUNCTION__, '1.3.4', 'ProtectTheChildren_Helpers::processPosts');
-        ProtectTheChildren_Helpers::processPosts($post);
+        Helpers::processPosts($post);
     }
 }
